Add fallback tests
